"Implement ApiResponse<T> class and update GetDepartmentAsync method in DepartmentController"
Description:
Implemented the ApiResponse<T> class to encapsulate API responses with consistent structures. Updated the GetDepartmentAsync method in the DepartmentController to utilize the ApiResponse<DepartmentModel> for providing standardized responses. Removed the manual setting of the IsError property and allowed it to be computed based on the presence of the Message property, ensuring accurate error indication in the response.